Move initialize() to opentrons.hardware_control.
